requirements.txt
spynner
pyserial
argparse
python-inotify
